Customer entry dialog is now fully functional. Records may be added to the table and the storage is happening flawlessly.
*NOTE*
I had some issues as each time that I added a new record, the previous records would be rewritten to the table as well as the new one. After quite a bit of frustration, I tried commenting out the file write and read code from the `CustomerCtl.addNew()` method. For some reason, I had placed the following into the `CustomerCtl.addNew()` function:

```java
    public void addNew(CustomerModel cust) {
        records.add(cust);
        row = getRecordCount() - 1;
        
        try {
            storeData();
            connect();
        } catch (DataStoreException ex) {
            entry.setMessage(ex.getMessage());
            entry.setSourceMethodName("Customers()");
            entry.setParameters(new Object[]{cust});
            entry.setThrown(ex);
            Starter.logger.error(entry);
        }
    }
```

I do not know why I hadn't just added the new record to the `ArrayList<CustomerModel>` object and not stored the data to the file, then read it back in. After commenting out the entire `try...catch` block, each record was only stored to the table once, regardless of how many customers I added in one display of the customer entry dialog. Therefore, I have deleted the entire `try...catch` block, since everything is now working as expected.

I've also added a task and window for adding brokers and/or agents into another table. Since this is very similar to the customer objects, I should be able to simply copy and paste the `CustomerModel` and `CustomerCtl` objects, renaming them as `BrokerModel` and `BrokerCtl`. Hopefully, this will save some time in development of the system.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>