Broker Entry Dialog Complete
`Brokers`, the broker/agent entry dialog is completed. None of the fields on this dialog are required, so I had to modify the `PostalCodeVerifier` and the `StateAbbrVerifier` classes to only perform verification if there is text entered into the `stateField`s and `zipField`s on the various dialogs and windows. Otherwise, I would have had to create two more `InputVerifier` classes in order to have one check only if the fields are required and the other to check only if there is actually data present in the field. I figured that this was the best, and easiest, way to handle this.

I am commencing to work on the `Booker` class, which is the load booking dialog.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>