BrokerSelector Dialog Complete
I have completed the `BrokerSelector` dialog and have it populating the broker/agent information on the `Booker` load booking window.

Other modifications include modifying the `BrokerCtl` class to include a `hasNext()` function to determine whether or not there are more records in the table while performing loops on the table's data.

Obviously, there are also the required changes to the `Booker` load booking window to show the `BrokerSelector` dialog and populate the broker/agent information on the `Booker` window.

Also, I have created the `CustomerSelector` dialog, though I am just beginning to work on it. I will make another commit and pull request once the `CustomerSelector` dialog is complete.

After I am done with the `CustomerSelector` dialog, I will then work on creating the `LoadModel` and `LoadCtl` classes for storing and reading the load information to the database.

All in all, I believe the application is coming along rather nicely. Of course, I have yet to begin working on the accounting portion of the system. I am sure that I will run into all kinds of problems once I being working on that. We will see, however...

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>