Updated VersionCalculator One Last Time
I have updated the `VersionCalculator` class to reflect the command-line switch that is discussed in Issue #9. I've tested these changes with both command-line switches and without the command-line switch. Everything is working as expected.

**NOTE**
Just remember to place either `--debugging` or `-d` on the Project Properties...Run...Arguments field to make sure that the build increases the version info.

***NOTE***
Because the version will be different on every computer that is working on this project, the versioning for the releases will be based upon the version that is showing on my computer. At the time of the release, I will hard-code the version into the places it needs to be in the application, such as on the status bar of the Load Master window and in the About box.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>