Enhancement Modifications
I have made some modifications to some of the classes and windows to provide for a little enhanced functionality. The modified classes are:

* `BrokerCtl`
* `CustomerCtl`
* `LoadCtl`
* `StopCtl`
* `Booker`
* `Brokers`
* `Customers`
* `LoadMaster`
* `Starter`

The modifications that I made were to keep the `fileProgress` progress bar visible at all times, because when I had it set to not being visible, then made it visible, the made it not visible, it messed up the appearance of the status bar and altered the size of the tips label for some reason. To work around that issue, I am just leaving it visible at all times.

Also, I have changed the controller classes to use the `fileProgress` progress bar to show the progress of opening the various table data files and reading in the records. To reset the `fileProgress`, I have reset its value to zero (0) in the `doCancel` event of each of the windows. This way, when the user completes, for example, adding a customer to the database, when the window closes, the `fileProgress` progress bar will have its value reset to zero.

Granted, at this time, the progress indicator value rises very rapidly and looks as if the progress indicator goes from zero to maximum value in an instance. But, as the datasets grow, the indicator will provide the user some feedback as to the progress of reading the data into the application. By having the `fileProgress` progress bar being utilized, the user will know whether the application is working on something, or if it has locked up for some unknown reason.

Just as an FYI, the `fileProgress` progress bar can be used for any time-consuming functionality to provide feedback to the user that the system is working.

***Programming Note:***
The `fileProgress` is currently only being used for opening the data files. I am going to also use it to show the progress of writing out the data files, but will have it run backwards from maximum value to zero. I simply like that idea because the indicator moves left to right while reading in the data, it should move right to left while writing the data back out to the file.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>