Minor Update to Postal Code Verifier
I made a minor update to the `PostalCodeVerifier` that determines whether a valid US Zip Code or Canadian Postal Code was entered into the `zipField`. The modification is to allow the word "Unavailable" to be entered into the field in the event the user does not have the postal code for the customer. "Unavailable" may be entered in any case and still pass the check.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>