Book Load Wizard
I am in the process of making the booking of a new load into a wizard, using the Wizard API. However, I am having issues with getting the Finish button to enable after a minimum of two stops have been selected into the stops table. I will create a detailed issue for this, but needed to commit the current changes to the repository in order to seek help in getting this wizard working.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>