Minor Modifications and Updates
### Modifications:

I modified two (2) methods in the controller objects. The changes are:

* `storeData` --> `close`
* `private void update([Object]Model model)` to `public void update([Object]Model model)`

I changed the names of two (2) methods in `LoadMaster` window:

* `doShowArrival` --> `doArrival`
* `doNewLoad` --> `doBookLoad`

I changed the arrival method name because the only visual cue regarding an arrival is an update to the load progress bar. Beyond that, there is nothing visual about an arrival at a stop.

I changed the load booking method name to better fit with the vernacular of the rest of the interface. Every place that a "new load" is mentioned within the interface, it is referred to as "booking a load," therefore I felt the method name should match, especially since the method shows the "Book Load Wizard".

### Updates

I have placed the storage logic behind the arrival and departure methods. However, we still need to create a departure dialog to receive the gathered information, such as bill of lading (BOL) number, number of pieces, weight, signed by, etc. This information is gathered from pickups and deliveries, though only signed by is required at the delivery. We need to figure a way to determine whether a stop is a pickup or a delivery, so that we can gather the appropriate information at that point.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>