One Departure Dialog Complete
I believe that I have the `DepartPickupDialog` complete, including all record saving/updating code in the `doShowDeparture` method of the `LoadMaster` window. However, I cannot test this code until I have a load to work with, which requires the Book Load Wizard (Issue #34)to be working. Once we get the Wizard working, I will be able to test this code out. ***I will create an issue to remind me to test this functionality, once the Wizard is working properly.***

I need to create the dialog and code for departing from a delivery yet, but that should only involve collecting the name of the person who signed for the freight.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>