Data Store Errors Fixed in the Book Load Wizard
Per comments from @jkovalsky, a Data Store Error occurred when he was navigating to the Broker pane of the Book Load Wizard (no issue filed, communicated via Facebook Messenger). The reason that he received that error is because he had no brokers listed in the `.loadmaster/data/brokers.tbl` data file.

This issue was one I had not contemplated when originally designing the Wizard, so it was good that it was pointed out to me. I have corrected this error by placing a check for the `[BrokerCtl].getRecordCount() == 0`. If there are no records in the brokers table, a `NewBrokerDlg` dialog is presented to the user to enter a broker into the table. Once a broker is entered into the table, the user is presented with the Wizard's Broker page, where they can simply click next or press enter.

Likewise, on the Stops page of the Wizard, I have added an item to the list as the second item, which is, "Add a new customer...". If the user selects this item, then the `NewCustomerEntryDlg` dialog is presented, allowing the customer to enter a new customer. Once the customer is entered, I am trying to make that the selected customer in the `CustomerSelector` dialog's customers list drop-down, but am not having any luck so far. Even so, the user can drop down the list and scroll all the way to the bottom and the new customer is there.

This functionality is still a little buggy, but it is getting better...

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>