Changes per Review
I have made the changes per the review by @jkovalsky in Pull Request #41.

I deleted the `About.main()` method, the commented out lines for creating a separator in the System tasks pane and setting the default button on the About dialog, since I am trapping the enter and escape keys to close the dialog.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>