Review Required
I found some issues on the Book New Load Wizard, so I started working on them and have now run into something that I do not understand.

Changes that I have made:

*  Deleted and recreated the `customerList` `JComboBox` on the customer selection dialog in the final step because, for some odd reason, the pop-up would not go away after the selection was made, until I clicked somewhere else on the form or pressed Escape.
*  Added the `validateContents()` function on the `BrokerPage` of the wizard, because the second step was not requiring that a broker or agent be selected, which is required in the database.
* I placed a data check around the parsing of the early and late times from the add stop dialog, as I want to verify that the time parts (hours and minutes) are valid and kept getting a `NullPointerException` when I would try to parse them after splitting the text at the colon (:). I am still getting the `NullPointerException` on the same line as before, even with a check for text in place.

New problem that popped up:

Now, when I get to the final step of the Wizard, adding stops, the Add Stop dialog will not even open. I have no clue what caused this issue to arise and need some serious assistance. Please help me figure out why this Wizard is not cooperating and working the way it should. It is, truly, holding up the rest of the project...

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>