Update to BrokerPage
I have updated the `BrokerPage` to allow the user to enter a new broker/agent to the database by selecting the `brokerList` entry, "Add New Broker/Agent..."

By allowing the user to do this, they will not need to use the `NewBrokerDlg` or `Brokers` internal frame prior to running the load booking wizard.

Also, I am working on adding a summary page as the last page of the wizard. By providing this page, the user will be able to review the data entered and go back if corrections need to be made. With this page added, the `StopsPage` will be set to `WizardController.MODE_CAN_CONTINUE` and the summary page will be set to `WizardController.MODE_CAN_FINISH`. 

_NOTE_: I thought about making the `StopsPage` allow finishing the wizard, but believe that we should show the summary and let the user finish from there. If you have other ideas, please let me know. This is not something fixed in stone, so we can discuss it freely.

As this is a weeknight, and I have to make a few deliveries tomorrow, I am going to stop here. I will pick up tomorrow with creating the entries in the `loads.tbl` and `stops.tbl` data files. Once this is done, the Book New Load Wizard will be complete and work can finally continue. As of right now, we (I) am running 7 days behind schedule for the first release.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>