Class Cleanup
I have gone through all of the visual component classes (`JFrame`, `JDialog`, `JInternalFrame`, etc) and removed the `main()` method that existed in any of them. In the `LoadMaster` and `SettingsDialog` classes, I added the comment header, as it was the NetBeans default, so now there are changelogs at the top of those two classes as well.

This commit is modest in what has been completed, but I wanted to get this cleanup taken care of before I started working on the Book Load Wizard to save the data collected to the data store files.

Since the changes in this commit were very minor, I am not requesting a review prior to pulling and merging into the master branch.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>