Adding Data Collected to the Wizard Data Map
I have performed some more cleanup on these classes and added logic(?) to store the collected data on each `WizardPage` to the wizard data `Map`. Also, I have modified the table layout of the `stopsTable` to better coincide with the changes made by @jkovalsky, and added a column for the Customer ID value.

I am seeking input on the logic of storing the Customer IDs to the wizard data map prior to attempting to store the data out to the data store files.

Thank you in advance for checking my logic for me!

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>