Creating Book Load Wizard Summary Page
I am working on displaying the data the user provided during the course of the Book Load Wizard steps on a Summary Page, using a `JEditorPane` to show the information well-formatted in HTML. However, when I run the application and get to this step of the wizard, the data map has no data in it. 

On each step of the wizard, I am using the `putWizardData()` method to save the data the user provides to the `wizardDataMap` object. I am providing the following keys for each piece of data:

- order: Order Number
- trip: Trip Number
- truck.pay: Gross Pay
- miles: Loaded Miles
- commodity: Product being hauled
- hazmat: HazMat CheckBox
- tarped: Tarped CheckBox
- team: Team CheckBox
- twic: TWIC CheckBox
- top.customer: Top Customer CheckBox
- cbd: Signature and Tally CheckBox
- ramps: Ramps CheckBox
- brokerID: Broker/Agent Unique ID - parsed from the `brokerList.getSelectedItem()` method
- stop#: One Entry for each stop

However, when I set a breakpoint at the beginning of the `SummaryPage` constructor and check the length of the object returned from the `getWizardDataMap()` method call, the size is 0. 

I need help figuring out why there is no data in the wizard data map when I have explicitly set the data to it for every field that I am interested in storing.

*NOTE*: I have not started attempting to write the data out to the data files yet, because I can not even seem to read the data simply to show the information to the user, so writing to file will, most likely, not work yet either.

I discovered this issue because I am getting an NPE displayed in the `summaryEditor` instead of the HTML that I am creating. The NPE is occurring on line 98 of the `SummaryPage` class, which is the first location that the data is being accessed.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>