Minor Modifications for Cleanup
I have added the trademark symbol (™) to the Project and Vendor Names, as both of these are registered trademarks of PekinSOFT™ Systems. Also added the copyright and trademark information to the bottom of the project description for the About box, as well as adding the Project Name to the About box titlebar.

One last modification is that I changed the titlebar text of the main window to pull the Project Name from the properties file.

Since all of these are minor modifications, I will simply create a pull request and merge them without any review being required.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>