Proposal Accepted
I have played with the Proposed Summary Page Solution (PR #48) and made a couple minor modifications, as the broker email was showing in the summary report as a text field when there was no email. I was checking for the length of the email to be greater than zero, however, I forgot that for the `String.split("~")` function to work properly, a blank space had to be saved to the data file if no data was provided for a field. I have that fixed to where if there is no email address, the email field will not be printed to the report.

Also, I had to play around to get the stops to show up in the summary report and believe that I have it working now. There may need to be some tweaks that will be made over time to improve this summary report, but I believe that I will now be able to start working on saving the data to the data files.

Once again, thank you @jkaovalsy for your help with the NPE that I was getting. I do not understand how I keep writing code that causes them, and you are so good at finding them. I guess that is what makes us a good team! ;-)

I am going to push my modifications to this branch and merge to the master, as I believe it is ready to do so.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>