Book Load Wizard ALMOST Complete
I have gotten everything working in the Book Load Wizard, except that it is not saving the stops to the stops table file (see all attached images). I've looked at the code in the `StopCtl.close()` method and it all seems to be correct. Perhaps a fresh set of eyes will see what I must be missing.

When I leave the `stop.tbl` file open in Notepad++, when I switch back to it after running the Wizard, it prompts me to reload the file because it was modified by another program. When I reload the file, it is still devoid of data (see the final image).

Please advise on what I have missed in the `StopCtl` class that is preventing the data from being stored in the `.loadmaster/data/stops.tbl` file.

The load data is saving perfectly (see the fifth image). This is why I am confused as to why the stops are not being saved out to the `stops.tbl` file. 

Also, I am getting no errors and am attaching the log file to this push.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>