Book Load Wizard ALMOST Complete Part 2
OK, so now all data is writing out to the files properly. However, I am getting an NPE in the `LoadCtl` class at line 484, where I am trying to load the stops that belong to the load. I am positive this is some logic flaw on my part, so anything that you can point out to me would be helpful. Here is the code in question:

```java
479        if ( stops != null ) {
480            boolean exists = false;
481            
482            for ( StopModel stop : stops.getList() ) { // All stops in database
483                for ( LoadModel load : records ) { // All loads in database
484                    for ( StopModel s : load.getStops() ) { // All stops in load
485                        if ( s.getTripNumber().equals(load.getTrip()) ) {
486                            exists = true;
487                        }
488                    } // All stops in load
489                } // All loads in database
490                
491                if ( !exists ) {
492                    stops.addNew(stop);
493                }
494            } // All stops in database
```

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>