Book Load Wizard Complete!
The Book Load Wizard is ***complete***! The data is storing properly, to the appropriate data files, as expected. I found out that the NPE that I was getting before was because I had not initialized the `stops` list in the `LoadModel` class. Once I was initializing that list, no more NPE. 

It seems to me that the Wizard is working 100% as expected, but I would like a review by @jkovalsky, just to be sure.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>