Chart of Accounts Created
Since we have gotten the Book New Load Wizard working properly, we need to start working on the accounting side of the project. The reason being that we need to have an Accounts Receivable account into which we can place the gross pay from the booked load.

For this purpose, I have created a default chart of accounts to be used with the system. We may add the ability to add a custom chart of accounts to the system at a later version. For now, the default chart of accounts is valid for every one-truck owner/operator and breaks expenses down sufficiently. I have also added five (5) asset accounts: 

* Checking Account
* Savings Account
* Retirement Savings Account
* Fuel Card
* Accounts Receivable

Furthermore, I have added one (1) liability account, Accounts Payable. These twenty-four (24) accounts will suffice for 99+% of our target market for this version of Load Master.

***NOTE***: I will not merge this branch until the Wizard branch is merged to the master.

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>