Load Booking Wizard
I have added the additional functionality of prompting the user to start the trip just entered, in the event that they are not already on a trip. In other words, if the "load.current" project setting is "No Active Load", then a question is asked of the user to see if they would like to start the trip they just entered. If the user responds positively, then information is stored to the settings file, including:

* Trip #
* Stop count
* Zero (0) as the current stop

Furthermore, there is still one more thing to do to have this wizard completely finished, and that is to write the gross revenue to the Accounts Payable file. I have not started working on the accounting functionality to the point where this is doable yet, but will create an Issue as a reminder that this still needs to be done. 

Other than connecting to the accounting system, the Book New Load Wizard is complete!

-SC
Signed-off-by:Sean Carrick <PekinSOFT@outlook.com>