Book Load Wizard with Accounting
I now have the Book Load Wizard complete with storing data to the Accounts Receivable account, since that is where the revenue needs to be stored until the load is complete.

This functionality involved creating a new dialog, `StartTripDialog`, as well as the following new classes:

* `EntryModel` : This is the definition of a General Ledger transaction entry.
* `EntryCtl` : This is the controller for creating, deleting, reading, and writing the General Ledger file. ***NOTE*** this is not being used yet, as the AR account journal entries never get written to the GL, themselves. They will only be written to the GL as the driver settles the load, using all of the money on-hand in the AR account to pay out to other accounts, such as insurance, fuel, maintenance, wages, etc.
* `ReceivablesModel` : This is the definition of a receivable.
* `ReceivablesCtl` : This is the controller for creating, deleting, reading, and writing the receivables to the receivables journal file.

Now, when the user completes the Book Load Wizard, they will be prompted to start the trip, if `Starter.props.getProperty("load.current")` is "No Active Load". If there is an active load, this methodology will not execute when the driver books the load.

Instead, we need to create a dialog from which the driver can view all booked loads, in order that they pick up from their shipper. From that dialog, the driver will be able to select a load and start the trip.

Regardless, the project seems to be coming along nicely!

-SC