Fuel Entry Functionality
I have created the `FuelPurchaseDialog`, `FuelPurchaseModel` and `FuelPurchaseCtl` classes for allowing a fuel purchase to be entered into the system. However, these classes have not yet been tested, as I need to provide a method of moving money from the gross truck pay in the Accounts Receivable account to the Fuel Card account before a fuel purchase can be entered into the GL and Fuel Journal. Looking at the code, I believe that this functionality will work splendidly, barring any NPEs, which keep showing up to annoy me. ;-)

Anyway, I am committing this to the `scarrick/FuelEntry` branch at this time and am going to switch to working on some modifications in existing, functional tools, such as limiting the Broker/Agent and Customer lists by state, for example. Some of those tables are beginning to grow very long and a user will not want to have to scroll through, possibly, thousands of records to find the one they are looking for. Also, I need to create some sort of logic to make sure the user cannot double-enter a broker/agent or customer that may already exist in the database.

-SC