Broker Searchability Working
I have updated the `BrokerCtl` to allow for the searching of records based upon city and state fields. Now, the list may be limited to only those brokers in a given city or state. I have this functionality working on the `BrokerSelector` dialog, but I just discovered something that I should have known...The `BrokerSelector` is not used anywhere at this time. In the Book Load Wizard, their is a broker selection list directly on the `BrokerPage` and the `BrokerSelector` is not displayed from there. Placing the required controls on the `BrokerPage` of the wizard worries me, as I am afraid of breaking it, due to how hard it was to get it working. Therefore, I am making a command decision at this time to not modify the `BrokersPage` of the Book Load Wizard. We will need to do this in the future, but for now I am satisfied that the functionality exists for brokers to limit the list(s) that may use those records. I am pushing this commit up to the repository, as is. I will request @jkovalsky to review it, and merge it to the master if it works effectively for him.

***Note***
@jkovalsky, set your Project Properties to use the `BrokerSelector` class as the startup class for testing purposes. Once the testing is complete and is working as expected, please go into the `BrokerCtl` class an uncomment the following lines:

In the `connect()` method:

```java
        BufferedReader in;
//        LoadMaster.fileProgress.setMinimum(0);
//        LoadMaster.fileProgress.setMaximum(
//                Starter.props.getPropertyAsInt("table.brokers.records", "0"));
//        LoadMaster.fileProgress.setVisible(true);
        
        try {
```

Each of the lines referencing `LoadMaster.fileProgress` need to be uncommented prior to pushing the pull request to the master branch. Thank you!

-SC