Customer Searchability Working
I have updated the `CustomerCtl` to allow for the searching of records based upon company, city, and state fields. Now, the list may be limited to only those customers in a given city or state, or from a given company. I have this functionality working on the `CustomerSelector` dialog. I will request @jkovalsky to review it, and merge it to the master if it works effectively for him.

Thank you!

-SC