Fuel Entry and Cash Advance
I have gotten the Fuel Entry well on its way, but I cannot test it quite yet, as I am still working on the Cash Advance feature. However, I do have the Cash Advance to the Fuel Card functional, I have kept going with the modifications that I was working on.

To that end, let me explain what those modifications are. I have created an abstract class in a new package called `com.pekinsoft.loadmaster.api`. The abstract class that I created there is called `AbstractJournal`. I decided to create this class because I was rewriting the same code again and again with my Journal controller classes. Therefore, I decided that I should create a class that did a lot of what I was redoing and allow me to only be required to override those methods that dealt with the files:

* `connect`
* `close`
* `save`
* `createAndAddRecord`
* `buildRecordLine`
* `createFileIfNecessary`

Because my `AbstractJournal` is using `Object`s, I have also created method overrides for:

* `first`
* `previous`
* `next`
* `last`
* `get()`
* `get(int idx)`

The only reason that I am overriding those methods is to cast the `Object` used by the `AbstractJournal` into the `*Model` for which the extending class is created. This way, we won't need to be doing a lot of casting around the project, though there is still quite a bit of that going on.

At this point, I am successfully writing to the General Ledger, the Fuel Journal, and the Reserves Journal. I will slowly but surely modify some of the other Journal controller classes that I have already written to extend the `AbstractJournal` class.

I have also gotten the batch method of posting to the General Ledger functional. There are now two new settings on the Accounting tab of the Application Settings Dialog:

* A `JCheckBox` for choosing the batch posting method, which will require the user to manually post journal entries to the General Ledger (though I have not created that accounting task yet).
* A `JSpinner` that allows the user to set the value from 1 to 100, inclusive, as the percentage of the gross revenue they are allowed to take on the active trip.

For that second bullet point, I have the `CashAdvanceDialog` working off of that `JSpinner` and not allowing the user to take any advances beyond what is allowed by that setting. For example, if the `maxAdvanceAmount` is set at `30` and the gross revenue on the load is $1,000.00, they are prevented from taking any more than $300.00 in advances. As the `CashAdvanceDialog` loads, it pulls in all advances taken from the Fuel Card Journal and totals them up. It then subtracts that amount from the percentage of the gross for the load that is stored in the Receivables Journal. Once they have hit there limit, the `availableAmountField` shows zero (0), or what they have left available, and they cannot exceed that amount. It took some work, but I got it working like a champ.

On a sadder note, all of a sudden, something broke in the Summary Page of the Book Load Wizard. The summary no longer shows the added stops for the load, though they are still being stored out to the stops table. I will create an issue for that, as well as for creating a way of filtering the Brokers List on the Brokers page of the Wizard. I have figured out how to make that work, but have been leery of breaking the Wizard again. I will get to those two items soon enough, however.

Anyway, I managed to accomplish quite a bit in the past 38 hours of marathon coding. What can I say? I hit the zone and could not stop. I was working on these various house-keeping tasks, looked up and it was 05:00 this morning. Needless to say, I am now sufficiently beat that I am ready to go to bed and it is only 2020 in the evening.

As Porky the Pig says, "Thea, Thea, That's All Folks!" I am wiped! Goodnight!

-SC