Code Cleanup of Controller Objects Complete
As part of this round of code cleanup, I turned `AbstractJournal` into a *generic* class. By doing this, I was able to remove all of the casting that I was doing in the other controller classes, since the previous version of `AbstractJournal` was using `Object`s, instead of the actual `*Model` classes. Now, each controller is declared as:

```java
public class FuelCardCtl extends AbstractJournal<FuelCardModel> {
   // ...
}
```

Also, in the modification of `AbstractJournal`, all but one (1) of the `public` methods are concrete. The only exception is `public boolean postTransactions()`, which needs to be implemented by the extending classes.

There are also two (2) `protected` methods that need to have their implementation created in the extending classes:

* `buildRecordLine(T model)` to build the record line for storing to the data file from the getters of the `T` type class.
* `createAndAddRecord(String[] line)` for taking the freshly read in and split `java.lang.String` array and placing the data into the `T` type class.

I thought about using reflection for this, which would mean that all three (3) of those `abstract` methods would not need to be `abstract`. However, I do not know how to determine which data to use with each setter via reflection, as the `Method methods[] = <class>.getMethods()` functionality is not guaranteed to return the reflected methods in any particular order. And, as far as I can determine, the methods will not necessarily be returned in the same order with successive invocations. If there was a guarantee that the methods would always come from the `<class>.getMethods()` implementation in the same exact order, we could use reflection. If we were able to use reflection, we would not need all of the controller classes in the `com.pekinsoft.loadmaster.controller` package. Instead, we would simply need to create instances of the `com.pekinsoft.loadmaster.api.AbstractJournal` class. At this time, however, I have `AbstractJournal` as abstract as I am able to get it.

The rest of this cleanup process was updating all of the controllers to reflect the changes to `AbstractJournal` and have the controllers extend it. Furthermore, I even set the database tables and the `EntryCtl` for the General Ledger to extend `AbstractJournal`, even though the database table and General Ledger controllers do not need to implement the `postTransactions` method. Therefore, in those controllers, I have simply put `throw new UnsupportedOperationException("Not necessary for this object.")`. The reason I set the message of that exception the way I did is because that operation will ***never*** be supported for those controllers. If there is a way to make an `abstract` method definition optional for the extending classes, I am all ears!

Once I complete this cleanup session, I had zero (0) errors anywhere, so even though the implementation in the controllers changed internally, the external interfaces are still valid. Strike one up for **proper** object-oriented programming practices!

From this point on, whenever a controller class is created for any model, the `AbstractJournal` is available to be used to keep the controller's code nice and clean.

-SC