Broker List Filtering Completed-ish
I have the Book Load Wizard's Broker Information page list able to be filtered by:

* All - Default
* State
* City
* Company
* Contact
* Phone

However, Contact and Phone filtering is currently not available, as contact filtering does not seem to be working and I need to think about filtering by phone number for a while. Other than that, filtering is now possible to narrow down the list as it grows through use of the system.

This commit fixes Issue #59.

-SC