Abstracting and Genericizing Controller
Along the lines of my `scarrick/FuelEntry` branch, I am performing code cleaning. Why? Simply put, I broke something somewhere along the line, after my `scarrick/Accounting` branch push.

I discovered this because of the testing system that @jkovalsky pushed. Every branch of mine **after** `scarrick/Accounting` causes the tests to fail. Therefore, I brought my local copy back to my accounting branch and created this branch off of it.

So, what I have done is to create two new classes: `AbstractJournal` and `JournalInterface`. I will discuss these classes in detail below. Let me just say that I have genericized the controller to the point where we will only need to use that one controller, and not a different controller for every model that needs access to the file system. However, I will discuss the `JournalInterface` class first.

The `JournalInterface` class is just what its name implies: an interface for journals to be used by their respective models. Each model that needs file system access for data storage, and wants to use the `AbstractJournal` class, ***must*** implement this interface. `JournalInterface` exposes two (2) methods at this time: `load(String[] data)` and `String buildRecordLine()`. Let us take a look at these two methods that must be implemented by models wanting to use `AbstractJournal` as their controller.

`load(String[] data)` accepts a `java.lang.String` array containing the data values for the individual fields of a record. The model's implementation takes each element of the `data` array and places it into the appropriate member field. Since the model is taking care of building the string that will be placed into the data file, the model knows the order of the data in the `data` array and can, therefore, place the correct data into the proper member field.

`String buildRecordLine()` allows the model to take the data from its member fields and place them into a `java.lang.String` that it can return back to the calling object, which will most likely be the `AbstractJournal`, as I cannot foresee any other object needing the model's data in this manner.

Now, let us look at the `AbstractJournal` class. This class is using generics in that the `ArrayList` and `record` objects within it are declared as type `JournalInterface`, so any class implementing that interface can use the `AbstractJournal` for reading and writing its data file. However, the names of some of the methods in `AbstractJournal` will not make sense outside of the context of an accounting system.

Within `AbstractJournal`, I am using some reflection to allow me to create new objects of the type that is currently using the class. To that end, the constructor of `AbstractJournal` takes two parameters, like this:

```java
    public AbstractJournal (JournalInterface model, String filePath)
            throws DataStoreException
```

The first parameter accepts any class that implements the `JournalInterface`. Therefore, access to its class is available with this class. So, I was able to use the constructor of
the class that implements the interface in order to not have an entire `ArrayList` filled with the same record multiple times. I create new instances of the model objects by invoking this method:

```java
    private Constructor getConstructor() {
        Constructor cons = null;
        
        try {
            cons = record.getClass().getDeclaredConstructor(String.class);
            cons.setAccessible(true);        } catch ( Exception ex ) {
            entry.setMessage(ex.getLocalizedMessage());
            entry.setThrown(ex);
            entry.setSourceMethodName("getConstructor");
            entry.setLevel(Level.WARNING);
            entry.setParameters(null);
            entry.setInstant(Instant.now());
            Starter.logger.error(entry);
        }
        
        return cons;
    }
```

To be able to initialize, or re-initialize, the single record object, simply call this method like so: `record = (JournalInterface)getConstructor().newInstance(lines);`. I have not tried doing this without casting to a `JournalInterface`, but as far as I understand things, the cast is not actually required.

I have deleted the `FuelPurchaseCtl` class from `com.pekinsoft.loadmaster.controller`, as I have updated the `FuelPurchaseDialog` to use the `AbstractJournal` class instead. All I needed to do was change one line in the `FuelPurchaseDialog` where I declared the local variable `records` to be of type `AbstractJournal`, then call the journal's constructor. Nothing else needed changed, which I take as a good sign.

Once I made these changes, I ran the automated tests and all three passed, so I have not broken anything yet. I will continue to work on modifying the controllers and models until I can completely remove the `com.pekinsoft.loadmaster.controller` package.

One last thing to mention, I have generated documentation using `doxygen` and will upload the `LoadMasterDoxyFile` to the root of the project separately. If you have `doxygen` on your system, you can download the `LoadMasterDoxyFile` and change the paths in it to suit your system to generate documentation on your local repository.

That is all I have for now.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>